+++ src/utils/Integrator.java
@@ -16,63 +16,63 @@ public class Integrator {
         return result;
     }
 
-    public static float[] gearPredictorCorrector(int order, Integrable model, int step, float deltaT) {
+    public static double[] gearPredictorCorrector(int order, Integrable model, int step, double deltaT) {
         //r_function
         FunctionWithDerivatives r_function = model.getrFunction();
         //force_function
-        BiFunction<Float, Float, Float> force_function = model.getForceFunction();
+        BiFunction<Double, Double, Double> force_function = model.getForceFunction();
         //mass
-        float mass = model.getMass();
+        double mass = model.getMass();
 
-        float[][] coefficents = {
+        double[][] coefficents = {
                 {},
                 {},
-                {0f, 1f, 1f},
-                {1f / 6f, 5f / 6f, 1f, 1f / 3f},
-                {19f / 90f, 3f / 4f, 1f, 1f / 2f, 1f / 12f},
-                {3f / 16f, 251f / 360f, 1f, 11f / 18f, 1f / 6f, 1f / 60f}
+                {0.0, 1.0, 1.0},
+                {1.0 / 6.0, 5.0 / 6.0, 1.0, 1.0 / 3.0},
+                {19.0 / 90.0, 3.0 / 4.0, 1.0, 1.0 / 2.0, 1.0 / 12.0},
+                {3.0 / 16.0, 251.0 / 360.0, 1.0, 11.0 / 18.0, 1.0 / 6.0, 1.0 / 60.0}
         };
 
         // 1) Predict
-        float[] rp = new float[order + 1];
+        double[] rp = new double[order + 1];
         for (int i = 0; i <= order; i++) {
             for (int j = i; j <= order; j++) {
-                rp[i] += r_function.calculateDerivative(j, step - 1) * (float) Math.pow(deltaT, j - i) / factorial(j - i);
+                rp[i] += r_function.calculateDerivative(j, step - 1) * (double) Math.pow(deltaT, j - i) / factorial(j - i);
             }
         }
 
         // 2) Evaluate
 
         // deltaA = a(t+deltaT) - ap(t+deltaT)
-        float deltaA = force_function.apply(rp[0], rp[1]) / mass - rp[2];
-        float deltaR2 = deltaA * deltaT * deltaT / 2.0f;
+        double deltaA = force_function.apply(rp[0], rp[1]) / mass - rp[2];
+        double deltaR2 = deltaA * deltaT * deltaT / 2.0;
 
         // 3) Correct
-        float[] r = new float[order + 1];
+        double[] r = new double[order + 1];
 
         for (int i = 0; i <= order; i++) { //Only r and v
-            r[i] = rp[i] + coefficents[order][i] * deltaR2 * factorial(i) / (float) Math.pow(deltaT, i);
+            r[i] = rp[i] + coefficents[order][i] * deltaR2 * factorial(i) / (double) Math.pow(deltaT, i);
         }
 
         return r;
     }
 
-    public static float[] beeman(Integrable model, int step, float deltaT) {
+    public static double[] beeman(Integrable model, int step, double deltaT) {
         //r_function
         FunctionWithDerivatives r_function = model.getrFunction();
         //force_function
-        BiFunction<Float, Float, Float> force_function = model.getForceFunction();
+        BiFunction<Double, Double, Double> force_function = model.getForceFunction();
         //mass
-        float mass = model.getMass();
+        double mass = model.getMass();
 
-        float[] r = new float[2]; // r and v
+        double[] r = new double[2]; // r and v
 
         r[0] = r_function.calculateDerivative(0, step - 1)
                 + r_function.calculateDerivative(1, step - 1) * deltaT
                 + 2f / 3f * r_function.calculateDerivative(2, step - 1) * deltaT * deltaT
                 - 1f / 6f * r_function.calculateDerivative(2, step - 2) * deltaT * deltaT;
 
-        float vp = r_function.calculateDerivative(1, step - 1)
+        double vp = r_function.calculateDerivative(1, step - 1)
                 + 3f / 2f * r_function.calculateDerivative(2, step - 1) * deltaT
                 - 1f / 2f * r_function.calculateDerivative(2, step - 2) * deltaT;
 
@@ -84,15 +84,15 @@ public class Integrator {
         return r;
     }
 
-    public static float[] originalVerlet(Integrable model, int step, float deltaT) {
+    public static double[] originalVerlet(Integrable model, int step, double deltaT) {
         //r_function
         FunctionWithDerivatives r_function = model.getrFunction();
         //force_function
-        BiFunction<Float, Float, Float> force_function = model.getForceFunction();
+        BiFunction<Double, Double, Double> force_function = model.getForceFunction();
         //mass
-        float mass = model.getMass();
+        double mass = model.getMass();
 
-        float[] r = new float[2]; // r and v
+        double[] r = new double[2]; // r and v
 
         r[0] = 2*r_function.calculateDerivative(0, step - 1)
                 - r_function.calculateDerivative(0, step - 2)
